# swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0",
        "title": "Survey Service",
        "description": "Microservice for designing and administering surveys.",
        "contact": {
            "name": "Amida Technology Solutions",
            "url": "http://www.amida.com",
            "email": "hello@amida-tech.com"
        },
        "license": {
            "name": "Apache 2.0"
        }
    },
    "basePath": "/api/v1.0",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/answers": {
            "x-swagger-router-controller": "answer.controller",
            "post": {
                "description": "This persists answers to survey questions.  Old answers are soft deleted.  For each question for which there is no answer property, any existing answer will be soft deleted.",
                "summary": "Creates answers",
                "operationId": "createAnswers",
                "parameters": [{
                    "name": "surveyAnswers",
                    "in": "body",
                    "description": "Answers to the survey",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/surveyAnswers"
                    }
                }],
                "responses": {
                    "204": {
                        "description": "Successful response."
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            },
            "get": {
                "summary": "Gets answers",
                "operationId": "getAnswers",
                "parameters": [{
                    "name": "survey-id",
                    "in": "query",
                    "description": "Survey id",
                    "required": true,
                    "type": "integer"
                }],
                "responses": {
                    "200": {
                        "description": "Successful response.",
                        "schema": {
                            "$ref": "#/definitions/linkedAnswers"
                        }
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/answers/csv": {
            "x-swagger-router-controller": "answer.controller",
            "get": {
                "summary": "Exports answers as csv file",
                "operationId": "exportAnswers",
                "produces": ["text/csv", "application/json"],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            },
            "post": {
                "summary": "Imports answer csv file",
                "operationId": "importAnswers",
                "consumes": ["multipart/form-data"],
                "parameters": [{
                    "name": "answercsv",
                    "in": "formData",
                    "description": "Answer CSV file",
                    "required": true,
                    "type": "file"
                }, {
                    "name": "questionidmap",
                    "in": "formData",
                    "description": "Question Id Map",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "surveyidmap",
                    "in": "formData",
                    "description": "Question Id Map",
                    "required": true,
                    "type": "string"
                }],
                "responses": {
                    "204": {
                        "description": "Successful response"
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/answers/multi-user-csv": {
            "x-swagger-router-controller": "answer.controller",
            "get": {
                "summary": "Exports answers as csv file",
                "operationId": "exportMultiUserAnswers",
                "produces": ["text/csv", "application/json"],
                "parameters": [{
                    "name": "user-ids",
                    "in": "query",
                    "required": true,
                    "type": "array",
                    "items": {
                        "type": "integer",
                        "minimum": 1
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            },
            "post": {
                "summary": "Imports answer csv file",
                "operationId": "importAnswers",
                "consumes": ["multipart/form-data"],
                "parameters": [{
                    "name": "answercsv",
                    "in": "formData",
                    "description": "Answer CSV file",
                    "required": true,
                    "type": "file"
                }, {
                    "name": "questionidmap",
                    "in": "formData",
                    "description": "Question Id Map",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "surveyidmap",
                    "in": "formData",
                    "description": "Question Id Map",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "useridmap",
                    "in": "formData",
                    "description": "User Id Map",
                    "required": false,
                    "type": "string"
                }],
                "responses": {
                    "204": {
                        "description": "Successful response"
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/answers/export": {
            "x-swagger-router-controller": "answer.controller",
            "get": {
                "summary": "List answers export ready",
                "operationId": "listAnswersExport",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/answers/multi-user-export": {
            "x-swagger-router-controller": "answer.controller",
            "get": {
                "summary": "List answers multi user export ready",
                "operationId": "listAnswersMultiUserExport",
                "parameters": [{
                    "name": "user-ids",
                    "in": "query",
                    "required": true,
                    "type": "array",
                    "items": {
                        "type": "integer",
                        "minimum": 1
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/answers/queries": {
            "x-swagger-router-controller": "answer.controller",
            "post": {
                "summary": "Searches all answers and returns count of matching users",
                "operationId": "searchAnswers",
                "parameters": [{
                    "name": "query",
                    "in": "body",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/answersQuery"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Successful response.",
                        "schema": {
                            "type": "object"
                        }
                    },
                   "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/answers/identifier-queries": {
            "x-swagger-router-controller": "answer.controller",
            "post": {
                "summary": "Searches all answers and returns count of matching users using identifiers",
                "operationId": "countParticipantsIdentifiers",
                "parameters": [{
                    "name": "query",
                    "in": "body",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/answerIdentifiersQuery"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Successful response.",
                        "schema": {
                            "type": "object"
                        }
                    },
                   "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/answers/user-ids": {
            "x-swagger-router-controller": "answer.controller",
            "post": {
                "summary": "Searches all answers and returns user ids",
                "operationId": "searchParticipants",
                "parameters": [{
                    "name": "query",
                    "in": "body",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/answersQuery"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Successful response.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "required": ["userId"],
                                "properties": {
                                    "userId": {
                                        "type": "integer",
                                        "minimum": 1
                                    }
                                },
                                "additionalProperties": false
                            }
                        }
                    },
                   "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/answers/identifier-user-ids": {
            "x-swagger-router-controller": "answer.controller",
            "post": {
                "summary": "Searches all answers and returns user ids",
                "operationId": "searchParticipantsIdentifiers",
                "parameters": [{
                    "name": "query",
                    "in": "body",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/answerIdentifiersQuery"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Successful response.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "required": ["userId"],
                                "properties": {
                                    "userId": {
                                        "type": "integer",
                                        "minimum": 1
                                    }
                                },
                                "additionalProperties": false
                            }
                        }
                    },
                   "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/answers/federated": {
            "x-swagger-router-controller": "answer.controller",
            "post": {
                "summary": "Searches all answers and returns federated answers",
                "operationId": "federatedListAnswers",
                "parameters": [{
                    "name": "query",
                    "in": "body",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/answerIdentifiersQuery"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Successful response.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
        "/answers/federated-queries": {
            "x-swagger-router-controller": "answer.controller",
            "post": {
                "summary": "Searches all answers and returns count of matching users",
                "operationId": "federatedSearchAnswers",
                "parameters": [{
                    "name": "query",
                    "in": "body",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/answerIdentifiersQuery"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Successful response.",
                        "schema": {
                            "type": "object"
                        }
                    },
                   "default": {
                        "$ref": "#/responses/genericError"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "self": {
            "type": "apiKey",
            "in": "header",
            "name": "Authorization"
        }
    },
    "responses": {
        "genericError": {
            "description": "Error response",
            "schema": {
                "$ref": "#/definitions/error"
            }
        },
...
    },
    "definitions": {
...
        "error": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "answersQuery": {
            "type": "object",
            "required": ["questions"],
            "properties": {
                "questions": {
                    "$ref": "#/definitions/questionQueries"
                },
                "federated": {
                    "type": "boolean"
                }
            },
            "additionalProperties": false
        },
        "questionQuery": {
            "type": "object",
            "required": ["id", "answers"],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "exclude": {
                    "type": "boolean"
                },
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/filterAnswer"
                    },
                    "minItems": 1
                }
            },
            "additionalProperties": false
        },
        "answerIdentifiersQuery": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/answerIdentifierQuery"
            }
        },
        "questionQueries": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/questionQuery"
            }
        },
...
        "choicesItem": {
            "type": "object",
            "required": ["id"],
            "properties": {
                "id": {
                    "type": "integer",
                    "minimum": 1
                },
                "boolValue": {
                    "type": "boolean"
                },
                "integerValue": {
                    "type": "integer"
                },
                "floatValue": {
                    "type": "number"
                },
                "textValue": {
                    "type": "string"
                },
                "yearValue": {
                    "type": "string",
                    "pattern": "^\\d{4}$"
                },
                "monthValue": {
                    "type": "string",
                    "pattern": "^\\d{2}$"
                },
                "dayValue": {
                    "type": "string",
                    "pattern": "^\\d{2}$"
                },
                "zipcodeValue": {
                    "type": "string",
                    "pattern": "^\\d{5}$"
                }
            },
            "maxProperties": 2
        },
        "answer": {
            "type": "object",
            "properties": {
                "multipleIndex": {
                    "type": "integer",
                    "minimum": 0
                },
                "choice": {
                    "type": "integer",
                    "minimum": 1
                },
                "choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/choicesItem"
                    },
                    "minItems": 1
                },
                "textValue": {
                    "type": "string"
                },
                "boolValue": {
                    "type": "boolean"
                },
                "dateValue": {
                    "type": "string",
                    "format": "date"
                },
                "numberValue": {
                    "type": "number"
                },
                "integerValue": {
                    "type": "integer"
                },
                "floatValue": {
                    "type": "number"
                },
                "feetInchesValue": {
                    "type": "object",
                    "properties": {
                        "feet": {
                            "type": "integer"
                        },
                        "inches": {
                            "type": "integer"
                        }
                    },
                    "minProperties": 1,
                    "additionalProperties": false
                },
                "bloodPressureValue": {
                    "type": "object",
                    "required": ["systolic", "diastolic"],
                    "properties": {
                        "systolic": {
                            "type": "integer"
                        },
                        "diastolic": {
                            "type": "integer"
                        }
                    },
                    "additionalProperties": false
                },
                "fileValue": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "name": {
                            "type": "string",
                            "minLength": 1
                        },
                        "content": {
                            "type": "string",
                            "minLength": 1,
                            "description": "Base64 encoded binary or text content"
                        }
                    },
                    "minProperties": 1,
                    "maxProperties": 2
                },
                "yearValue": {
                    "type": "string",
                    "pattern": "^\\d{4}$"
                },
                "monthValue": {
                    "type": "string",
                    "pattern": "^\\d{2}$"
                },
                "dayValue": {
                    "type": "string",
                    "pattern": "^\\d{2}$"
                },
                "zipcodeValue": {
                    "type": "string",
                    "pattern": "^\\d{5}$"
                }
            },
            "minProperties": 1,
            "maxProperties": 2,
            "additionalProperties": false
        },
        "filterAnswer": {
            "type": "object",
            "properties": {
                "choice": {
                    "type": "integer",
                    "minimum": 1
                },
                "textValue": {
                    "type": "string"
                },
                "boolValue": {
                    "type": "boolean"
                },
                "dateValue": {
                    "type": "string",
                    "format": "date"
                },
                "numberValue": {
                    "type": "number"
                },
                "integerValue": {
                    "type": "integer"
                },
                "integerRange": {
                    "type": "object",
                    "properties": {
                        "min": {
                            "type": "integer"
                        },
                        "max": {
                            "type": "integer"
                        }
                    },
                    "minProperties": 1,
                    "additionalProperties": false
                },
                "floatValue": {
                    "type": "number"
                },
                "feetInchesValue": {
                    "type": "object",
                    "properties": {
                        "feet": {
                            "type": "integer"
                        },
                        "inches": {
                            "type": "integer"
                        }
                    },
                    "minProperties": 1,
                    "additionalProperties": false
                },
                "bloodPressureValue": {
                    "type": "object",
                    "required": ["systolic", "diastolic"],
                    "properties": {
                        "systolic": {
                            "type": "integer"
                        },
                        "diastolic": {
                            "type": "integer"
                        }
                    },
                    "additionalProperties": false
                },
                "yearValue": {
                    "type": "string",
                    "pattern": "^\\d{4}$"
                },
                "monthValue": {
                    "type": "string",
                    "pattern": "^\\d{2}$"
                },
                "dayValue": {
                    "type": "string",
                    "pattern": "^\\d{2}$"
                },
                "zipcodeValue": {
                    "type":"string",
                    "pattern": "^\\d{5}$"
                }
            },
            "minProperties": 1,
            "maxProperties": 2,
            "additionalProperties": false
        },
...
        "answerIdentifierQuery": {
            "type": "object",
            "properties": {
                "identifier": {
                    "type": "string",
                    "minLength": 1
                },
                "questionText": {
                    "type": "string",
                    "minLength": 1
                },
                "exclude": {
                    "type": "boolean"
                },
                "questionChoiceText": {
                    "type": "string",
                    "minLength": 1
                },
                "textValue": {
                    "type": "string"
                },
                "boolValue": {
                    "type": "boolean"
                },
                "dateValue": {
                    "type": "string",
                    "format": "date"
                },
                "numberValue": {
                    "type": "number"
                },
                "integerValue": {
                    "type": "integer"
                },
                "integerRange": {
                    "type": "object",
                    "properties": {
                        "min": {
                            "type": "integer"
                        },
                        "max": {
                            "type": "integer"
                        }
                    },
                    "minProperties": 1,
                    "additionalProperties": false
                },
                "floatValue": {
                    "type": "number"
                },
                "feetInchesValue": {
                    "type": "object",
                    "properties": {
                        "feet": {
                            "type": "integer"
                        },
                        "inches": {
                            "type": "integer"
                        }
                    },
                    "minProperties": 1,
                    "additionalProperties": false
                },
                "bloodPressureValue": {
                    "type": "object",
                    "required": ["systolic", "diastolic"],
                    "properties": {
                        "systolic": {
                            "type": "integer"
                        },
                        "diastolic": {
                            "type": "integer"
                        }
                    },
                    "additionalProperties": false
                },
                "yearValue": {
                    "type": "string",
                    "pattern": "^\\d{4}$"
                },
                "monthValue": {
                    "type": "string",
                    "pattern": "^\\d{2}$"
                },
                "dayValue": {
                    "type": "string",
                    "pattern": "^\\d{2}$"
                }
            },
            "minProperties": 2,
            "maxProperties": 5,
            "additionalProperties": false
        },
...
        "linkedAnswer": {
            "type": "object",
            "required": [
                "questionId"
            ],
            "properties": {
                "questionId": {
                    "type": "integer",
                    "minimum": 1
                },
                "language": {
                    "type": "string",
                    "minLength": 2
                },
                "meta": {
                    "type": "object"
                },
                "answer": {
                    "$ref": "#/definitions/answer"
                },
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/answer"
                    }
                }
            },
            "additionalProperties": false
        },
        "linkedAnswers": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/linkedAnswer"
            }
        },
        "surveyAnswers": {
            "type": "object",
            "required": [
                "surveyId",
                "answers"
            ],
            "properties": {
                "surveyId": {
                    "type": "integer",
                    "minimum": 1
                },
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/linkedAnswer"
                    },
                    "minItems": 1
                },
                "language": {
                    "type": "string",
                    "minLength": 2
                }
            },
            "additionalProperties": false
        },
...
    }
}